Updates addon to version 0.1.1
Updates the addon version number in the CHANGELOG and config file.

Improves user handling in the Dockerfile by checking if the `gotify` user already exists before attempting to create it. It also includes fallback mechanisms for user creation if `adduser` is not available, ensuring compatibility across different base images.

Enhances the `run.sh` script to properly set data directory permissions and execute the Gotify binary, even if the `gotify` user doesn't exist or `su-exec` or `gosu` are not available. This ensures the addon can run with or without a dedicated user and across different environments.